+++ yandex/data_source_yandex_mdb_postgresql_cluster.go
@@ -250,6 +250,10 @@ func dataSourceYandexMDBPostgreSQLCluster() *schema.Resource {
 								},
 							},
 						},
+						"conn_limit": {
+							Type:     schema.TypeInt,
+							Optional: true,
+						},
 					},
 				},
 			},