+++ yandex/resource_yandex_mdb_postgresql_cluster_test.go
@@ -311,6 +311,9 @@ func testAccCheckMDBPGClusterHasUsers(r string, perms map[string][]string) resou
 var defaultUserSettings = map[string]interface{}{
 	"conn_limit": int64(50),
 }
+var testAccMDBPGClusterConfigUpdatedCheckConnLimitMap = map[string]int64{
+	"alice": 42,
+}
 
 func testAccCheckUnmodifiedUserSettings(r string) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
@@ -332,9 +335,16 @@ func testAccCheckUnmodifiedUserSettings(r string) resource.TestCheckFunc {
 		if err != nil {
 			return err
 		}
+
+		defaultConnLimit := defaultUserSettings["conn_limit"].(int64)
 		for _, user := range resp.Users {
-			defaultConnLimit := defaultUserSettings["conn_limit"].(int64)
-			if user.ConnLimit != defaultConnLimit {
+			v, ok := testAccMDBPGClusterConfigUpdatedCheckConnLimitMap[user.Name]
+			if ok {
+				if user.ConnLimit != v {
+					return fmt.Errorf("Field 'conn_limit' wasn`t changed for user %s with value %d ",
+						user.Name, user.ConnLimit)
+				}
+			} else if user.ConnLimit != defaultConnLimit {
 				return fmt.Errorf("Unmodified field 'conn_limit' was changed for user %s with value %d ",
 					user.Name, user.ConnLimit)
 			}
@@ -586,6 +596,8 @@ resource "yandex_mdb_postgresql_cluster" "foo" {
   user {
     name     = "alice"
     password = "mysecurepassword"
+
+	conn_limit = 42
 
     permission {
       database_name = "testdb"